Remove h5py patching and unused xy projection from odin factories
The h5py patching was needed for DetectorProjection which is no longer used.
DetectorLogicalDownsampler works directly with the unpatched geometry file
by renaming dim_0/dim_1 to x/y in the fold transform.

Prompt: Can we now remove the _xy_projection for Odin? iirc it was added as
a workaround to get ROI support, but I am not sure.