Fix correlation histogram reconfiguration losing axis sources
When reconfiguring a correlation histogram plot, SpecBasedConfigurationStep.on_enter
was creating a PlotterSelection from the initial config but not copying axis_sources
into correlation_axes. This caused the orchestrator to use the standard pipeline
(MergingStreamAssembler) instead of the correlation pipeline, resulting in the
plotter receiving a dict instead of CorrelationHistogramData.

Prompt: Getting error in @log.txt when reconfiguring a correlation plot (2d),
or reloading the page (sometimes). Please think and find out what is going on.